Move page models to hook file
